Part 4
I created the schematic for Circuit1 (mostly because it seemed the easiest).  There is a question and discussion in Piazza that I followed.  Please make sure that the comments in the pdf are accessible.  They have the information about the Estimated Critical Path for each output and Actual Critical Path for the circuit.
https://piazza.com/class/j6f4twcbksp6dk?cid=74
https://piazza.com/class/j6f4twcbksp6dk?cid=71